- début modification format données d'échange PC/FRDM car trop lourd

+++ controLab/frdmcontrolview (ABDULLATIF Mouhamadi's conflicted copy 2013-07-12).h
@@ -35,7 +35,7 @@ public:
 
 private slots:
     void onRessourceListChanged(const QString &);
-    void onSetPointValueChanged(int);
+    void onSetPointValueChanged(double);
     void onSetPointSliderLevelReleased();
 
 Q_SIGNALS:
